refactor: Move calendar header components to new directory
Move `WeekHeaderNavigation`, `SwipableWeekHeader`, and
`WeekHeaderContent` components into a new
`calendarHeader` directory for better organization.  The
`Calendar` component now accepts a callback to update the
current date in the parent component.  A month navigation
button has been removed from the calendar.
